+++ NEWS.md
@@ -1,4 +1,4 @@
-# eegUtils 0.1.15.9000
+# eegUtils 0.1.15.9001
 
 ### Function changes
 - `as.data.frame.eeg_epochs()` now has a `cond_labels` parameter to select epochs with specific events and add the event label as an additional column.
@@ -7,6 +7,7 @@
 
 ### Internal changes / bug fixes
 - `plot_butterfly()` some `dplyr` use removed.
+- `run_ica()` refactored SOBI method, JADE dependency removed.
 - `montage_check()` command parses montage info when passed to `electrode_locations()`
 - `label_check()` added to help parse event labels
 - `proc_events()` added to help parse event labels during `select_epochs()` calls